anomaly-worker/requirements.txt
pika==1.3.2
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
pandas==2.3.3
scikit-learn==1.8.0
joblib==1.5.3
